+++ docs/level-7-use-custom-jinja2-filter-test-n-global/custom-jj2-plugin/filter.py
@@ -1,0 +1,13 @@
+import sys
+import base64
+from moban.extensions import JinjaFilter
+
+
+@JinjaFilter()
+def base64encode(string):
+    if sys.version_info[0] > 2:
+        content = base64.b64encode(string.encode('utf-8'))
+        content = content.decode('utf-8')
+    else:
+        content = base64.b64encode(string)
+    return content