+++ docs/level-7-use-custom-jinja2-filter-test-n-global/custom-jj2-plugin/test.py
@@ -1,0 +1,6 @@
+from moban.extensions import JinjaTest
+
+
+@JinjaTest()
+def level7(value):
+    return value == 'level7'