+++ moban/filters/repr.py
@@ -1,0 +1,9 @@
+from moban.extensions import JinjaFilter
+
+
+@JinjaFilter()
+def repr(string):
+    if isinstance(string, list):
+        return ["'{0}'".format(str(element)) for element in string]
+    else:
+        return "'{0}'".format(str(string))