+++ moban/mobanfile.py
@@ -1,7 +1,13 @@
 import os
+import re
+import sys
+from collections import defaultdict
+
+from lml.utils import do_import
+
 import moban.constants as constants
 import moban.reporter as reporter
-from moban.engine import EngineFactory
+from moban.engine import ENGINES
 from moban.utils import merge, parse_targets
 from moban.copier import Copier
 
@@ -23,19 +29,41 @@ def handle_copy(template_dirs, copy_config):
 
 
 def handle_targets(merged_options, targets):
-    list_of_templating_parameters = parse_targets(
-        merged_options, targets,
-    )
-    engine_class = EngineFactory.get_engine(
-        merged_options[constants.LABEL_TEMPLATE_TYPE]
-    )
-    engine = engine_class(
-        merged_options[constants.LABEL_TMPL_DIRS],
-        merged_options[constants.LABEL_CONFIG_DIR],
-    )
-    engine.render_to_files(list_of_templating_parameters)
-    engine.report()
-    return engine.number_of_templated_files()
+    list_of_templating_parameters = parse_targets(merged_options, targets)
+    jobs_for_each_engine = defaultdict(list)
+    for file_list in list_of_templating_parameters:
+        _, extension = os.path.splitext(file_list[0])
+        template_type = extension[1:]
+        if template_type not in ENGINES.all_types():
+            template_type = merged_options[constants.LABEL_TEMPLATE_TYPE]
+        jobs_for_each_engine[template_type].append(file_list)
+
+    count = 0
+    for template_type in jobs_for_each_engine.keys():
+        engine_class = ENGINES.get_engine(
+            template_type
+        )
+        engine = engine_class(
+            merged_options[constants.LABEL_TMPL_DIRS],
+            merged_options[constants.LABEL_CONFIG_DIR],
+        )
+        engine.render_to_files(jobs_for_each_engine[template_type])
+        engine.report()
+        count = count + engine.number_of_templated_files()
+    return count
+
+
+def handle_plugin_dirs(plugin_dirs):
+    for plugin_dir in plugin_dirs:
+        plugin_path = os.path.dirname(os.path.abspath(plugin_dir))
+        if plugin_path not in sys.path:
+            sys.path.append(plugin_path)
+        pysearchre = re.compile(".py$", re.IGNORECASE)
+        pluginfiles = filter(pysearchre.search, os.listdir(plugin_dir))
+        plugins = list(map(lambda fp: os.path.splitext(fp)[0], pluginfiles))
+        for plugin in plugins:
+            plugin_module = os.path.basename(plugin_dir) + "." + plugin
+            do_import(plugin_module)
 
 
 def handle_moban_file_v1(moban_file_configurations, command_line_options):
@@ -46,6 +74,9 @@ def handle_moban_file_v1(moban_file_configurations, command_line_options):
             moban_file_configurations[constants.LABEL_CONFIG],
         )
     merged_options = merge(command_line_options, constants.DEFAULT_OPTIONS)
+    plugins_dirs = merged_options.get(constants.LABEL_PLUGIN_DIRS)
+    if plugins_dirs:
+        handle_plugin_dirs(plugins_dirs)
 
     targets = moban_file_configurations.get(constants.LABEL_TARGETS)
     if targets:
@@ -56,7 +87,7 @@ def handle_moban_file_v1(moban_file_configurations, command_line_options):
     if constants.LABEL_COPY in moban_file_configurations:
         number_of_copied_files = handle_copy(
             merged_options[constants.LABEL_TMPL_DIRS],
-            moban_file_configurations[constants.LABEL_COPY]
+            moban_file_configurations[constants.LABEL_COPY],
         )
     else:
         number_of_copied_files = 0