+++ tests/test_hash_store.py
@@ -5,7 +5,7 @@ from moban.hashstore import HashStore
 class TestHashStore:
 
     def setUp(self):
-        self.source_template = os.path.join("tests", "fixtures", "a.template")
+        self.source_template = os.path.join("tests", "fixtures", "a.jj2")
         self.fixture = (
             "test.out", "test content".encode("utf-8"), self.source_template
         )