+++ tests/test_moban_file.py
@@ -7,6 +7,8 @@ class TestFinder:
     def setUp(self):
         self.patcher = patch("os.path.exists")
         self.fake_file_existence = self.patcher.start()
+        self.fake_file_existence.__name__ = 'fake'
+        self.fake_file_existence.__module__ = 'fake'
 
     def tearDown(self):
         self.patcher.stop()