+++ tests/test_utils.py
@@ -1,23 +1,11 @@
 import os
 import stat
 
-from nose.tools import raises, eq_
-from moban.utils import load_external_engine
+from nose.tools import eq_
+
 from moban.utils import file_permissions_copy
 from moban.utils import write_file_out
 from moban.utils import strip_off_trailing_new_lines
-from moban.engine import Engine
-
-
-@raises(ImportError)
-def test_load_external_engine():
-    load_external_engine("unknown_template_type")
-
-
-def test_load_mako_engine():  # fake mako
-    module = load_external_engine("mako")
-    engine_class = module.get_engine("mako")
-    assert engine_class == Engine
 
 
 def create_file(test_file, permission):