+++ CHANGELOG.rst
@@ -1,6 +1,26 @@
 Change log
 ================================================================================
 
+0.2.0 - 11-06-2018
+--------------------------------------------------------------------------------
+
+Added
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+#. `#18 <https://github.com/moremoban/moban/issues/18>`_: file exists test
+#. `#23 <https://github.com/moremoban/moban/issues/23>`_: custom jinja plugins
+#. `#26 <https://github.com/moremoban/moban/issues/26>`_: repr filter
+#. `#47 <https://github.com/moremoban/moban/issues/47>`_: allow the expansion of
+   template engine
+#. `#58 <https://github.com/moremoban/moban/issues/58>`_: allow template type
+   per template
+
+Updated
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+#. `#34 <https://github.com/moremoban/moban/issues/34>`_: fix plural message if
+   single file is processed
+
 0.1.4 - 29-May-2018
 --------------------------------------------------------------------------------